+++ src/main/java/ru/kuznetsov/bikeService/models/service/Usable.java
@@ -3,6 +3,6 @@ package ru.kuznetsov.bikeService.models.service;
 import ru.kuznetsov.bikeService.models.Showable;
 
 public interface Usable extends Showable {
-    int getManufacturer();
+    Long getManufacturer();
     String getModel();
 }
